+++ libdd-profiling/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- reduce OOM edges in `ProfileExporter::build`
+
+### Changed
+- Updated `prost` dependency from 0.13.5 to 0.14.1
+
 ## 1.0.0 - 2025-11-17
 
 Initial release.